ObservableRepositoryInvitationsClient

+++ Octokit.Reactive/Clients/IObservableRepositoryInvitationsClient.cs
@@ -13,7 +13,7 @@ namespace Octokit.Reactive
         /// See the <a href="https://developer.github.com/v3/repos/invitations/#accept-a-repository-invitation">API documentation</a> for more information.
         /// </remarks>        
         /// <param name="invitationId">The id of the invitation.</param>        
-        IObservable<bool> Accept(int invitationId);
+        IObservable<bool> Accept(long invitationId);
 
         /// <summary>
         /// Decline a repository invitation.
@@ -22,7 +22,7 @@ namespace Octokit.Reactive
         /// See the <a href="https://developer.github.com/v3/repos/invitations/#decline-a-repository-invitation">API documentation</a> for more information.
         /// </remarks>        
         /// <param name="invitationId">The id of the invitation.</param>        
-        IObservable<bool> Decline(int invitationId);
+        IObservable<bool> Decline(long invitationId);
 
         /// <summary>
         /// Deletes a repository invitation.
@@ -32,7 +32,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The id of the repository.</param>
         /// <param name="invitationId">The id of the invitation.</param>        
-        IObservable<bool> Delete(long repositoryId, int invitationId);
+        IObservable<bool> Delete(long repositoryId, long invitationId);
 
         /// <summary>
         /// Gets all invitations for the current user.
@@ -82,6 +82,6 @@ namespace Octokit.Reactive
         /// <param name="invitationId">The id of the invitation.</param>   
         /// <param name="permissions">The permission to set.</param>
         /// <returns><see cref="RepositoryInvitation"/></returns>
-        IObservable<RepositoryInvitation> Edit(long repositoryId, int invitationId, InvitationUpdate permissions);
+        IObservable<RepositoryInvitation> Edit(long repositoryId, long invitationId, InvitationUpdate permissions);
     }
 }
